extract assertion methods to _utils
